abstract network manipulation
