+++ lib/update.js
@@ -1,10 +1,20 @@
 "use strict"
 
+// UPDATE
+
 exports.data = (store, node, data) => {
-    // create/update data on node
+    // node must exists..
+    store.getObject(node);
+    // get data, compare with hash
+    // exists and different:
+    // - remove hash and add new data triples
+    // doesnt exists:
+    // - add new data triples
+    // exists and equal:
+    // - ignore
 };
 exports.name = (store, node, name) => {
-    // create/update name on node
+    store.getName(node);
 };
 exports.out = (store, set, out, node) => {
     // set node delete previous
